@@ -3,9 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^6.1.2",
+    "@craco/craco": "^7.1.0",
     "@tailwindcss/forms": "^0.3.2",
     "@tailwindcss/postcss": "^4.1.11",
+    "@tanstack/react-query": "^5.84.2",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
@@ -19,13 +20,13 @@
     "postcss-loader": "6.2.1",
     "postcss-nested": "^7.0.2",
     "postcss-preset-env": "^10.2.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-feather": "^2.0.9",
     "react-hook-form": "^7.3.5",
-    "react-query": "^3.15.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-query": "3.39.3",
+    "react-router-dom": "^6",
+    "react-scripts": "5.0.1",
     "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
   },
@@ -65,13 +66,14 @@
   },
   "proxy": "http://localhost:5000",
   "devDependencies": {
+    "@types/minimatch": "^6.0.0",
     "@types/react-router-dom": "^5.1.7",
-    "autoprefixer": "^10.4.21",
+    "autoprefixer": "^9",
     "eslint-plugin-prettier": "^3.4.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
-    "postcss": "8.4.31",
+    "postcss": "^7",
     "prettier": "^2.2.1",
-    "tailwindcss": "^4.1.11"
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat"
   },
   "resolutions": {
     "postcss": "8.4.31"
