@@ -1,16 +1,16 @@
 import './styles/index.css';
 
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import React from 'react';
-import ReactDOM from 'react-dom';
-import { QueryClient, QueryClientProvider } from 'react-query';
+import ReactDOM from 'react-dom/client';
 
 import App from './App';
 import { AuthenticationProvider } from './context/AuthenticationContext';
 import reportWebVitals from './reportWebVitals';
 
 const queryClient = new QueryClient();
-
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root')!);
+root.render(
   <React.StrictMode>
     <AuthenticationProvider>
       <QueryClientProvider client={queryClient}>
@@ -18,7 +18,6 @@ ReactDOM.render(
       </QueryClientProvider>
     </AuthenticationProvider>
   </React.StrictMode>,
-  document.getElementById('root'),
 );
 
 reportWebVitals();
