@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { Loader } from 'react-feather';
 import { useForm } from 'react-hook-form';
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 
 import useAuth from '../hooks/useAuth';
 import LoginRequest from '../models/auth/LoginRequest';
@@ -9,7 +9,7 @@ import authService from '../services/AuthService';
 
 export default function Login() {
   const { setAuthenticatedUser } = useAuth();
-  const history = useHistory();
+  const navigation = useNavigate();
 
   const [error, setError] = useState<string>();
 
@@ -23,7 +23,7 @@ export default function Login() {
     try {
       const data = await authService.login(loginRequest);
       setAuthenticatedUser(data.user);
-      history.push('/');
+      navigation('/');
     } catch (error) {
       setError(error.response.data.message);
     }
