@@ -1,12 +1,17 @@
-import { forwardRef, Module } from '@nestjs/common';
+import { Module, forwardRef } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { CourseModule } from '../course/course.module';
 import { ContentService } from './content.service';
+import { Content } from './content.entity';
+import { CourseModule } from 'src/course/course.module';
 
 @Module({
-  imports: [forwardRef(() => CourseModule)],
+  imports: [
+    TypeOrmModule.forFeature([Content]),
+    forwardRef(() => CourseModule),
+  ],
   controllers: [],
   providers: [ContentService],
-  exports: [ContentService],
+  exports: [ContentService, TypeOrmModule],
 })
 export class ContentModule {}
