@@ -1,13 +1,18 @@
-import { forwardRef, Module } from '@nestjs/common';
+import { Module, forwardRef } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { ContentModule } from '../content/content.module';
 import { CourseController } from './course.controller';
 import { CourseService } from './course.service';
+import { Course } from './course.entity';
+import { ContentModule } from 'src/content/content.module';
 
 @Module({
-  imports: [forwardRef(() => ContentModule)],
+  imports: [
+    TypeOrmModule.forFeature([Course]),
+    forwardRef(() => ContentModule),
+  ],
   controllers: [CourseController],
   providers: [CourseService],
-  exports: [CourseService],
+  exports: [CourseService, TypeOrmModule],
 })
 export class CourseModule {}
