@@ -1,11 +1,14 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
+import { User } from './user.entity';
 
 @Module({
+  imports: [TypeOrmModule.forFeature([User])], 
   controllers: [UserController],
   providers: [UserService],
-  exports: [UserService],
+  exports: [UserService, TypeOrmModule], 
 })
 export class UserModule {}
