@@ -25,14 +25,9 @@ class UserService {
     id: string,
     updateUserRequest: UpdateUserRequest,
   ): Promise<void> {
-    const {
-      firstName,
-      isActive,
-      lastName,
-      password,
-      role,
-      username,
-    } = updateUserRequest;
+    const { firstName, isActive, lastName, password, role, username } =
+      updateUserRequest;
+
     await apiService.put(`/api/users/${id}`, {
       firstName: firstName || undefined,
       lastName: lastName || undefined,
@@ -48,4 +43,5 @@ class UserService {
   }
 }
 
-export default new UserService();
+const UserServiceInstance = new UserService();
+export default UserServiceInstance;
