@@ -1,13 +1,8 @@
+/** @type {import('tailwindcss').Config} */
 module.exports = {
-  purge: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
-  darkMode: false,
+  content: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
   theme: {
     extend: {},
-  },
-  variants: {
-    extend: {
-      opacity: ['disabled'],
-    },
   },
   plugins: [require('@tailwindcss/forms')],
 };
